Time: 0 ms (100.00%) | Memory: 8.9 MB (7.24%) - LeetSync

+++ 50-powx-n/powx-n.cpp
@@ -3,7 +3,7 @@ public:
     double myPow(double x, int n) {
         long double y=1;
         
-        y=(long double)pow(x,n);
+        y=pow(x,n);
         
         
     return y;
